Bug fix - chance card health 'connection error'
When chance card health is picked by user and it's not possible to retrieve tweets live, a list of fallback keywords are to be displayed and the user should get to pick one from the list. Bug: when the user chooses a keyword the program gets stuck in a loop. Fixed: Fallback filename is now returned and passed into sentiment analysis.